+++ lib/lightning_web/router.ex
@@ -308,7 +308,12 @@ defmodule LightningWeb.Router do
   scope "/" do
     pipe_through [:browser, :require_authenticated_user, :require_superuser]
 
-    live_dashboard "/dashboard", metrics: LightningWeb.Telemetry
+    live_dashboard "/dashboard",
+      metrics: LightningWeb.Telemetry,
+      on_mount: [
+        {LightningWeb.InitAssigns, :default},
+        {LightningWeb.Hooks, :ensure_admin}
+      ]
   end
 
   do_in(:dev) do